Update Setting pages
