+++ packages/frontend/src/components/views/Settings.tsx
@@ -9,11 +9,10 @@ import { settingsPaths } from '$constants/settings-paths';
 import { Button } from '$elements/button/button';
 import { Heading } from '$elements/Heading';
 import { IconName } from '$elements/Icon';
-import { UpdatePageInfo } from '$renderers/seo/updatePageInfo';
 
-interface SettingsProps {
+type SettingsProps = {
   roles?: Role[];
-}
+};
 
 export const Settings: FC<SettingsProps> = ({ roles }) => {
   const urls = useMemo(
@@ -22,12 +21,12 @@ export const Settings: FC<SettingsProps> = ({ roles }) => {
         label: 'Account',
         items: [
           {
-            children: 'User preferences',
+            children: 'User Preferences',
             link: settingsPaths.userPreferences,
             icon: 'UserCircleIcon' as IconName,
           },
           {
-            children: 'Download my data',
+            children: 'Download My Data',
             link: settingsPaths.dataDownload,
             icon: 'CloudArrowDownIcon' as IconName,
           },
@@ -49,7 +48,7 @@ export const Settings: FC<SettingsProps> = ({ roles }) => {
           ...(roles?.includes(Role.TestUser)
             ? [
                 {
-                  children: 'Overrite data',
+                  children: 'Overwrite User Data',
                   link: settingsPaths.dataOverwrite,
                   icon: 'ExclamationTriangleIcon' as IconName,
                 },
@@ -62,9 +61,8 @@ export const Settings: FC<SettingsProps> = ({ roles }) => {
   );
 
   return (
-    <>
-      <UpdatePageInfo />
-      <section className="grid gap-8">
+    <section>
+      <div className="grid gap-8">
         {urls.map(({ label, items }) => (
           <List key={label} label={label}>
             {items.map((item) => (
@@ -76,13 +74,13 @@ export const Settings: FC<SettingsProps> = ({ roles }) => {
           <Heading>Theme</Heading>
           <ThemeSwitcher />
         </div>
-      </section>
+      </div>
       <div className="flex flex-col gap-2 mt-12 sm:flex-row">
         <HardRefreshButton />
         <Button accentColor="secondary" href="/auth/logout">
           Sign out
         </Button>
       </div>
-    </>
+    </section>
   );
 };
